requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-environ==0.10.0
Markdown==3.4.3
django-filter==23.1
rest-framework-roles==1.0.3
Pillow==9.4.0
django-tinymce==3.6.1
